requirements.txt
numpy>=1.21.0
pydantic>=2.0.0
pyyaml>=6.0.0
openai>=1.0.0
anthropic>=0.5.0
pytest>=7.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
scikit-learn>=1.0.0  # For similarity metrics 